separated CheckEditPreferences test, that's why added: CheckLogin and CheckPageSize tests Updated EditUserAvatar test
