+++ src/test/java/Test/EditUserAvatar.java
@@ -16,7 +16,7 @@ public class EditUserAvatar {
         String userAvatar = "//*[@id=\"user-avatar-picker\"]/div[1]";
 
         LoginSteps editUserAvatar = new UserAvatar();
-        editUserAvatar.loginToJira();
+        editUserAvatar.LoginToJira();
         ((UserAvatar) editUserAvatar).clickOnUserAvatar();
 
         assertEquals($(By.xpath(userAvatar)).getText(), "Select a User Avatar");
